Latin1 branch
Code to make source regex and destination be understood as latin1.
This causes undesired behavior because ascii/latin1/utf8 are encodings
and what we're replacing here are bytes. It makes much more sense to
convert the characters to unicode for both regex and destination and
the behavior will be more consistent